CMS-434: Move to `cms` dir
