+++ frontend/packages/component-library/package.json
@@ -149,6 +149,14 @@
       "import": "./dist/esm/cms/snapshot/index.mjs",
       "require": "./dist/cjs/cms/snapshot/index.js"
     },
+    "./cms/spacer": {
+      "types": {
+        "import": "./dist/esm/cms/spacer/index.d.ts",
+        "require": "./dist/cjs/cms/spacer/index.d.ts"
+      },
+      "import": "./dist/esm/cms/spacer/index.mjs",
+      "require": "./dist/cjs/cms/spacer/index.js"
+    },
     "./cms/tabGroup": {
       "types": {
         "import": "./dist/esm/cms/tabGroup/index.d.ts",
@@ -342,14 +350,6 @@
       },
       "import": "./dist/esm/slider/index.mjs",
       "require": "./dist/cjs/slider/index.js"
-    },
-    "./spacer": {
-      "types": {
-        "import": "./dist/esm/spacer/index.d.ts",
-        "require": "./dist/cjs/spacer/index.d.ts"
-      },
-      "import": "./dist/esm/spacer/index.mjs",
-      "require": "./dist/cjs/spacer/index.js"
     },
     "./tabs": {
       "types": {
