+++ frontend/packages/component-library/src/cms/spacer/README.md
@@ -1,0 +1,17 @@
+# `component-library/cms/spacer`
+
+## Consuming This Component
+
+This package exports one styled React component: [Spacer](Spacer.tsx).
+You can import it like this:
+
+```javascript
+import Spacer from '@code-dot-org/component-library/cms/spacer';
+```
+
+_Note:_ This component is used primarily for the Contentful CMS.
+
+For guidelines on how to use this component and the features it
+offers, [visit Storybook](https://code-dot-org.github.io/code-dot-org/component-library-storybook/).
+Or run storybook locally and go
+to [CMS / Spacer](http://localhost:6006/?path=/docs/cms-spacer--docs).
