--- frontend/packages/component-library/src/spacer/README.md
@@ -1,15 +1,0 @@
-# `component-library/spacer`
-
-## Consuming This Component
-
-This package exports one styled React component: [Spacer](Spacer.tsx).
-You can import it like this:
-
-```javascript
-import Spacer from '@code-dot-org/component-library/spacer';
-```
-
-For guidelines on how to use this component and the features it
-offers, [visit Storybook](https://code-dot-org.github.io/code-dot-org/component-library-storybook/?path=/docs/designsystem-spacer--docs).
-Or run storybook locally and go
-to [Design System / Spacer](http://localhost:6006/?path=/docs/designsystem-spacer--docs).
